fix: icons did not render on expo web
