+++ packages/common/src/dynamicLoading/dynamic-loading-setting.ts
@@ -1,67 +1,15 @@
 import { Platform } from 'react-native';
 
+import type { ExpoAssetModule, ExpoFontLoaderModule, ExpoFontUtilsModule } from './expo-global';
 import type { FontSource } from './types';
 
-type ExpoAssetModule = {
-  // definition from
-  // https://github.com/expo/expo/blob/1f5a5991d14aad09282d1ce1612b44d30e7e7d3d/packages/expo-asset/ios/AssetModule.swift#L23
-  downloadAsync: (uri: string, hash: string | undefined, type: string) => Promise<string>;
-};
+export type { LoadAsyncAsset } from './expo-global';
 
-// this is a file:// uri on native, or an object with uri and display on web
-export type LoadAsyncAsset = string | { uri: string; display: string };
-
-type ExpoFontLoaderModule = {
-  // definition from
-  // https://github.com/expo/expo/blob/1f5a5991d14aad09282d1ce1612b44d30e7e7d3d/packages/expo-font/ios/FontLoaderModule.swift#L18
-  getLoadedFonts: () => string[];
-  loadAsync: (fontFamilyAlias: string, asset: LoadAsyncAsset) => Promise<void>;
-};
-
-// RenderToImageResult needs to be usable as the `source` prop for image,
-// so it must stay compatible with ImageURISource type
-type RenderToImageResult = {
-  /**
-   * The file uri to the image.
-   */
-  uri: string;
-  /**
-   * Image width in dp.
-   */
-  width: number;
-  /**
-   * Image height in dp.
-   */
-  height: number;
-
-  /**
-   * Scale factor of the image. Multiply the dp dimensions by this value to get the dimensions in pixels.
-   * */
-  scale: number;
-};
-
-type ExpoFontUtilsModule = {
-  renderToImageAsync: (
-    glyph: string,
-    options: {
-      fontFamily: string;
-      size?: number;
-      lineHeight?: number;
-      color?: number;
-    },
-  ) => Promise<RenderToImageResult>;
-};
-
-declare global {
-  interface ExpoGlobal {
-    modules: {
-      ExpoAsset?: ExpoAssetModule;
-      ExpoFontLoader?: ExpoFontLoaderModule;
-      ExpoFontUtils?: ExpoFontUtilsModule;
-    };
-  }
-
-  var expo: ExpoGlobal | undefined;
+// requiring `expo-font` on web calls registerWebModule, thanks to which `getIsDynamicLoadingSupported` can return true on web
+if (Platform.OS === 'web' && globalThis.expo) {
+  try {
+    require('expo-font');
+  } catch (_err) {}
 }
 
 type ExpoGlobalType = {
@@ -75,11 +23,12 @@ type ExpoGlobalType = {
 function getIsDynamicLoadingSupported(globalObj: any): globalObj is {
   expo: ExpoGlobalType;
 } {
+  const expoModules = globalObj?.expo?.modules;
   return (
-    globalObj?.expo &&
-    (Platform.OS === 'web' || typeof globalObj.expo.modules?.ExpoAsset?.downloadAsync === 'function') &&
-    typeof globalObj.expo.modules?.ExpoFontLoader?.getLoadedFonts === 'function' &&
-    typeof globalObj.expo.modules?.ExpoFontLoader?.loadAsync === 'function'
+    !!expoModules &&
+    (Platform.OS === 'web' || typeof expoModules.ExpoAsset?.downloadAsync === 'function') &&
+    typeof expoModules.ExpoFontLoader?.getLoadedFonts === 'function' &&
+    typeof expoModules.ExpoFontLoader?.loadAsync === 'function'
   );
 }
 
@@ -91,7 +40,7 @@ export function getIsRenderToImageSupported(globalObj: any): globalObj is {
     };
   };
 } {
-  return globalObj?.expo && typeof globalObj.expo.modules?.ExpoFontUtils?.renderToImageAsync === 'function';
+  return typeof globalObj?.expo?.modules?.ExpoFontUtils?.renderToImageAsync === 'function';
 }
 
 export function assertExpoModulesPresent(globalObj: unknown): asserts globalObj is { expo: ExpoGlobalType } {
@@ -119,8 +68,9 @@ export const isDynamicLoadingSupported = () => getIsDynamicLoadingSupported(glob
 export const setDynamicLoadingEnabled = (value: boolean): boolean => {
   if (!getIsDynamicLoadingSupported(globalThis)) {
     if (process.env.NODE_ENV !== 'production' && !!value) {
+      const expoModules = globalThis.expo?.modules;
       const hasNecessaryExpoModules =
-        (Platform.OS === 'web' || !!globalThis.expo?.modules?.ExpoAsset) && !!globalThis.expo?.modules?.ExpoFontLoader;
+        (Platform.OS === 'web' || !!expoModules?.ExpoAsset) && !!expoModules?.ExpoFontLoader;
       const message = hasNecessaryExpoModules
         ? 'Expo is installed, but does not support dynamic font loading. Make sure to use Expo SDK 54 or newer.'
         : 'Necessary Expo modules not found. Dynamic font loading is not available when necessary Expo modules are not present.';
